Fixed 150+ compiler warnings (mostly switched to _s secure CRT funtions), hopefully didn't introduce new bugs by that. Fixed bug that could affect ability to use Memory DC Render Mode. (AfxHookGoldSrc/supportrender.cpp). Fixed bug in AfxCppCli/loader.cpp (_tfopen_s not being in binary mode), that could have caused problems for particular images.

+++ AfxHookGoldSrc/EasySampler.cpp
@@ -708,7 +708,7 @@ float EasySamplerSettings::FrameStrength_get() const
 {
 	return m_FrameStrength;
 }
-float EasySamplerSettings::Height_get() const
+int EasySamplerSettings::Height_get() const
 {
 	return m_Height;	
 }
